fix: suspense 적용을 위해 데이터 페칭로직 자식으로 위임
